refactor: split frontend components
